+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- handle error case of identifier being too large
 
 ## [v0.3.3]
 - encode error as header  [#10](https://github.com/xmidt-org/argus/pull/10)